requirements.txt
pginter>=0.2.3
Pillow~=9.5.0
